Add EKPUtils
