+++ src/ClimaCalibrate.jl
@@ -4,6 +4,20 @@ export project_dir
 
 project_dir() = dirname(Base.active_project())
 
+# TODO: Ask Nat about using Reexport later
+include("EKPUtils.jl")
+using .EKPUtils:
+    minibatcher_over_samples,
+    observation_series_from_samples,
+    g_ens_matrix,
+    get_metadata_for_nth_iteration,
+    get_observations_for_nth_iteration
+export minibatcher_over_samples,
+    observation_series_from_samples,
+    g_ens_matrix,
+    get_metadata_for_nth_iteration,
+    get_observations_for_nth_iteration
+
 include("ekp_interface.jl")
 include("model_interface.jl")
 include("slurm.jl")
