+++ test/observation_recipe.jl
@@ -1461,18 +1461,18 @@ end
         ),
     )
 
-    metadata1 = ObservationRecipe.get_metadata_for_nth_iteration(obs_series, 1)
-    metadata2 = ObservationRecipe.get_metadata_for_nth_iteration(obs_series, 2)
-    metadata3 = ObservationRecipe.get_metadata_for_nth_iteration(obs_series, 3)
+    metadata1 = ClimaCalibrate.get_metadata_for_nth_iteration(obs_series, 1)
+    metadata2 = ClimaCalibrate.get_metadata_for_nth_iteration(obs_series, 2)
+    metadata3 = ClimaCalibrate.get_metadata_for_nth_iteration(obs_series, 3)
     metadata_indices1 =
         ext._get_minibatch_indices_for_nth_iteration(obs_series, 1)
     metadata_indices2 =
         ext._get_minibatch_indices_for_nth_iteration(obs_series, 2)
     metadata_indices3 =
         ext._get_minibatch_indices_for_nth_iteration(obs_series, 3)
-    obs1 = ObservationRecipe.get_observations_for_nth_iteration(obs_series, 1)
-    obs2 = ObservationRecipe.get_observations_for_nth_iteration(obs_series, 2)
-    obs3 = ObservationRecipe.get_observations_for_nth_iteration(obs_series, 3)
+    obs1 = ClimaCalibrate.get_observations_for_nth_iteration(obs_series, 1)
+    obs2 = ClimaCalibrate.get_observations_for_nth_iteration(obs_series, 2)
+    obs3 = ClimaCalibrate.get_observations_for_nth_iteration(obs_series, 3)
 
     @test length(metadata1) == 3
     @test metadata1[1].attributes["short_name"] == "time"
